packages/core/pyproject.toml
[project]
name = "pyarnes-core"
version = "0.1.0"
description = "Core abstractions for the pyarnes agentic harness — types, errors, lifecycle, and logging."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [{ name = "Cognitivemesh" }]
keywords = ["agentic", "harness", "core", "types", "errors"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "loguru>=0.7",
    "funcy>=2.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-sdk>=1.0",
    "opentelemetry-exporter-otlp>=1.0",
]
ast = ["libcst>=1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyarnes_core"]
